# api/nodejs-server/api/swagger.yaml
swagger: "2.0"
info:
  description: "Provides a RESTful API for interacting with Smartsplit. Documentation is written in both French and English."
  version: "v1"
  title: "Smartsplit API"
  x-origin:
  - format: "yaml"
    url: "https://raw.githubusercontent.com/iptoki/smartsplit/develop/api/swagger.yaml"
  x-providerName: "smartsplit.org"
  x-logo:
    url: "http://smartsplit.org/wp-content/uploads/2019/05/Logo-v02-005-noir.png"
#host: api.smartsplit.org
basePath: "/v1"
...
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- jwt: []
paths:
...
  /rightHolders:
    get:
      tags:
      - "RightHolders"
      description: "Get a list of all right holder profiles"
      operationId: "getAllRightHolders"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolders"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
    post:
      tags:
      - "RightHolders"
      description: "This method creates a new right holder profile"
      operationId: "postRightHolder"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/rightHolder"
      responses:
        200:
          description: "OK. The right holder profile was created."
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}:
    get:
      tags:
      - "RightHolders"
      description: "Get a right holder's profile with the given ID"
      operationId: "getRightHolder"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
    put:
      tags:
      - "RightHolders"
      description: "This method updates a right holder profile"
      operationId: "updateRightHolder"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/rightHolder"
      responses:
        200:
          description: "OK. The right holder profile was created."
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
    delete:
      tags:
      - "RightHolders"
      description: "Delete a right holder's profile with the given ID"
      operationId: "deleteRightHolder"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      responses:
        200:
          description: "OK. Right holder profile deleted."
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/firstName:
    patch:
      tags:
      - "RightHolders"
      description: "Update right holder's first name with the given ID"
      operationId: "patchRightHolderFirstName"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "firstName"
        description: "The rights holder's first name"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/firstName"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              firstName:
                type: "string"
                example: "John"
                description: "The right holder's first name"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/lastName:
    patch:
      tags:
      - "RightHolders"
      description: "Update right holder's last name with the given ID"
      operationId: "patchRightHolderLastName"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "lastName"
        description: "The rights holder's last name"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/lastName"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              lastName:
                type: "string"
                example: "Smith"
                description: "The right holder's last name"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/editeur:
    patch:
      tags:
      - "RightHolders"
      description: "Mise à jour de l'état éditeur avec l'identifiant unique"
      operationId: "patchRightHolderEstEditeur"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The rights holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "editeur"
        description: "The rights holder's last name"
        required: true
        schema:
          $ref: "#/definitions/rightHolderEditeur"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              lastName:
                type: "boolean"
                example: true
                description: "Est un éditeur ou non"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/email:
    patch:
      tags:
      - "RightHolders"
      description: "Update right holder's email address with given ID"
      operationId: "patchRightHolderEmail"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "email"
        description: "The rights holder's email address"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/email"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              email:
                type: "string"
                example: "bob.andrews@example.com"
                description: "The right holder's email address"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
  /rightHolders/emailToRightHolderId:
    post:
      tags:
      - "RightHolders"
      description: "Get the right holder's ID given their email address"
      operationId: "postEmailToRightHolderId"
      parameters:
      - in: "body"
        name: "body"
        description: "request"
        required: true
        schema:
          $ref: "#/definitions/rightHolder"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              rightHolderId:
                type: "string"
                example: "134a8431-ea1f-4c04-ad1b-dfd27adee166"
                description: "The right holder's ID"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/requestSource:
    patch:
      tags:
      - "RightHolders"
      description: "Update right holder's requestSource (pochette or smartsplit) address with given ID"
      operationId: "patchRightHolderRequestSource"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "requestSource"
        description: "The rights holder's requestSource, either smartsplit or pochette upon request"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/requestSource"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              requestSource:
                type: "string"
                example: "Smartsplit"
                description: "The application requestSource of the right holder's last request"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      security:
      - jwt: []
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/ipi:
    patch:
      tags:
      - "RightHolders"
      description: "Update right holder's IPI number"
      operationId: "patchRightHolderIPI"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "ipi"
        description: "The right holder's IPI number"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/ipi"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              ipi:
                type: "string"
                example: "00004576"
                description: "The right holder's interested party information number"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/jurisdiction:
    patch:
      tags:
      - "RightHolders"
      description: "Update the jurisdiction for the given right holder"
      operationId: "patchRightHolderJurisdiction"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The artwork agreement's unique ID"
        required: true
        type: "string"
      - in: "body"
        name: "jurisdiction"
        description: "The jurisdiction of the given right holder"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/jurisdiction"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              jurisdiction:
                type: "string"
                example: "Canada"
                description: "The jusrisdiction of the artwork agreement of the given\
                  \ media"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/wallet:
    patch:
      tags:
      - "RightHolders"
      description: "Update the wallet address of a right holder"
      operationId: "patchRightHolderWallet"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "wallet"
        description: "The right holder's wallet address"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/wallet"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              wallet:
                type: "string"
                example: "0xdd87ae15f4be97e2739c9069ddef674f907d27a9"
                description: "The right holder's wallet address"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/artistName:
    patch:
      tags:
      - "RightHolders"
      description: "Update the artist name of a right holder"
      operationId: "patchRightHolderArtistName"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "artistName"
        description: "The right holder's artist name"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/artistName"
      responses:
        200:
          description: "OK"
          schema:
            properties:
              artistName:
                type: "string"
                example: "QuestLove"
                description: "The right holder's artist name"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/avatarImage:
    patch:
      tags:
      - "RightHolders"
      description: "Update the artist name of a right holder"
      operationId: "patchRightHolderAvatarImage"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "avatarImage"
        description: "The right holder's S3 Etag for the profile avatar image"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/avatarImage"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolder/properties/avatarImage"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/socialMediaLinks:
    patch:
      tags:
      - "RightHolders"
      description: "Update list of social media links for the given right holder"
      operationId: "patchRightHolderSocialMediaLinks"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "socialMediaLinks"
        description: "The object containing the given right holder profile's social media links"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/socialMediaLinks"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/groups:
    patch:
      tags:
      - "RightHolders"
      description: "Update the string set of groups for the given right holder"
      operationId: "patchRightHolderGroups"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "groups"
        description: "The array containing the given right holder profile's groups"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/groups"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
  /rightHolders/{rightHolderId}/defaultRoles:
    patch:
      tags:
      - "RightHolders"
      description: "Update default Roles for the given right holder"
      operationId: "patchRightHolderDefaultRoles"
      parameters:
      - name: "rightHolderId"
        in: "path"
        description: "The right holder's unique profile ID"
        required: true
        type: "string"
      - in: "body"
        name: "defaultRoles"
        description: "The array containing the given right holder profile's default Roles"
        required: true
        schema:
          $ref: "#/definitions/rightHolder/properties/defaultRoles"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/rightHolder"
        400:
          description: "Bad request"
        403:
          description: "Permission denied"
        404:
          description: "Endpoint not found"
        500:
          description: "An unknown error occured"
      x-swagger-router-controller: "RightHolders"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication"
    type: "basic"
  jwt:
    description: "A JSON Web Token (JWT) used for access control and authorization"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  rightHolders:
    type: "array"
    description: "An array of right holder profile information"
    items:
      $ref: "#/definitions/rightHolder"
  rightHolderEditeur:
    type: "object"
    properties:
      editeur:
        type: "boolean"
  rightHolder:
    type: "object"
    properties:
      rightHolderId:
        type: "string"
        description: "The unique Id for the right holder shared with AWS cognito user pools"
        example: "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g"
      firstName:
        type: "string"
        example: "John"
        description: "The first name displayed on the right holder's profile"
      lastName:
        type: "string"
        example: "Smith"
        description: "The last name displayed on the right holder's profile"
      email:
        type: "string"
        example: "john.smith@example.com"
        description: "The email address for this right holder profile"
      requestSource:
        type: "string"
        example: "pochette"
        description: "The last known application that the user made a request to update account/profile"
      jurisdiction:
        type: "string"
        example: "Canada"
        description: "The country that has jurisdiction over the rights of the right holder"
      ipi:
        type: "string"
        example: "00004576"
        description: "The right holder's interested party information number"
      wallet:
        type: "string"
        example: "0xdd87ae15f4be97e2739c9069ddef674f907d27a8"
        description: "The blockchain wallet address for this right holder profile"
      avatarImage:
        type: "string"
        description: "The file name key in AWS S3 for the given right holder avatar image"
        example: "image.jpg"
      defaultRoles:
        type: "array"
        description: "The array of the default roles for the given right holder profile" 
        items:
          $ref: "#/definitions/rightHolder/properties/defaultRoles"
      groups:
        type: "array"
        description: "The array of the groups for the given right holder profile" 
        items:
          $ref: "#/definitions/rightHolder/properties/groups"
      artistName:
        type: "string"
        description: "The artist name for the given right Holder"
        example: "Questlove"
      accountCreationType: 
        type: "string"
        description: "The string value that describes if the account is self registered or create by a split initiator"
        example: "initiatorCreatedUser"
      locale: 
        type: "string"
        description: "The string value of the locale detected from the browser upon user or initiator account creation"
        example: "fr"
      socialMediaLinks:
        type: "object"
        description: "The urls for the social media profiles for the given right holder, like Facebook, Instagram, Youtube, Twitter"
        additionalProperties:
          type: "string"
          description: "social media links for the given right holder profile"
      editeur:
        type: "boolean"
        description: "Est un ayant-droit éditeur"        
    example:
      rightHolderId: "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g"
      firstName: "John"
      lastName: "Smith"
      email: "john.smith@example.com"
      requestSource: "pochette"
      jurisdiction: "Canada"
      ipi: "00004576"
      wallet: "0xdd87ae15f4be97e2739c9069ddef674f907d27a8"
      avatarImage: "image.jpg"
      artistName: "Questlove"
      groups:
      - "Group 1"
      - "Group 2"
      defaultRoles:
      - "Songwriter"
      - "Producer"
      accountCreationType: "initiatorCreatedUser"
      locale: "en"
      editeur: true
      socialMediaLinks: 
        facebook: "https://facebook.com/ex"
        twitter: "https://twitter.com/ex"
        youtube: "https://youtube.com/ex"
...